Split UrlChooser and LocalFileChooser to new files
